Refactor usecases
