+++ src/domain/usecases/instance/GetCurrentUserUseCase.ts
@@ -1,7 +1,7 @@
-import { UseCase } from "../../CompositionRoot";
-import { FutureData } from "../entities/Future";
-import { User } from "../entities/User";
-import { InstanceRepository } from "../repositories/InstanceRepository";
+import { UseCase } from "../../../CompositionRoot";
+import { FutureData } from "../../entities/Future";
+import { User } from "../../entities/User";
+import { InstanceRepository } from "../../repositories/InstanceRepository";
 
 export class GetCurrentUserUseCase implements UseCase {
     constructor(private instanceRepository: InstanceRepository) {}
