+++ src/domain/usecases/instance/GetInstanceVersionUseCase.ts
@@ -1,6 +1,6 @@
-import { UseCase } from "../../CompositionRoot";
-import { FutureData } from "../entities/Future";
-import { InstanceRepository } from "../repositories/InstanceRepository";
+import { UseCase } from "../../../CompositionRoot";
+import { FutureData } from "../../entities/Future";
+import { InstanceRepository } from "../../repositories/InstanceRepository";
 
 export class GetInstanceVersionUseCase implements UseCase {
     constructor(private instanceRepository: InstanceRepository) {}
