+++ src/domain/usecases/instance/SearchUsersUseCase.ts
@@ -1,7 +1,7 @@
-import { UseCase } from "../../CompositionRoot";
-import { UserSearch } from "../../data/entities/SearchUser";
-import { FutureData } from "../entities/Future";
-import { InstanceRepository } from "../repositories/InstanceRepository";
+import { UseCase } from "../../../CompositionRoot";
+import { UserSearch } from "../../../data/entities/SearchUser";
+import { FutureData } from "../../entities/Future";
+import { InstanceRepository } from "../../repositories/InstanceRepository";
 
 export class SearchUsersUseCase implements UseCase {
     constructor(private instanceRepository: InstanceRepository) {}
